requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
colorama==0.4.6
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
distro==1.9.0
fonttools==4.55.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
openai==1.63.2
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
PyLaTeX==1.4.2
pyparsing==3.2.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
watchdog==6.0.0
weasyprint==64.0
webencodings==0.5.1
zopfli==0.2.3.post1
